Pass CARGO* and OUT_DIR env vars when running rustc.
I'm not sure how OUT_DIR was supposed to be getting set, since cargo
didn't seem to set it when we were acting as a RUSTC_WRAPPER. But it did
pass --out-dir as an argument, so I set OUT_DIR based on that. For the
CARGO* environment variables, I extracted their values from what was
passed by cargo to the rustc wrapper, returned them up to the main
program (by serializing as JSON, then deserializing) then set them
before we call into rustc code.

Fixes #9